% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sweep.R
\name{clean_sweep}
\alias{clean_sweep}
\title{Cleaning function to remove NA values from the original CSV file and save that as a new CSV}
\usage{
clean_sweep(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item the filepath to my original CSV
}}
}
\value{
a new CSV file with no NA values called new_data
}
\description{
Cleaning function to remove NA values from the original CSV file and save that as a new CSV
}
