% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_format_edit.R
\name{date_format}
\alias{date_format}
\title{Function to edit the date into separate columns}
\usage{
date_format()
}
\value{
one of the 2 new variables with the date separated into 3 columns
}
\description{
Function to edit the date into separate columns
}
