% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_barplot.R
\name{plotting_function}
\alias{plotting_function}
\title{Graphing function to display the most common insect families collected}
\usage{
plotting_function()
}
\value{
a bar plot showing the number of each family of insects that was collected
}
\description{
Graphing function to display the most common insect families collected
}
