% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_site_separation.R
\name{stream_separation}
\alias{stream_separation}
\title{Separates the stream sites in my original csv file and returns 2 separate new csv files}
\usage{
stream_separation()
}
\arguments{
\item{filename}{\itemize{
\item the filepath to my original CSV
}}
}
\value{
two new csv files, titled South mills and North mills
}
\description{
Separates the stream sites in my original csv file and returns 2 separate new csv files
}
